Bug fix: Exception while getting translations where 'from language' is specified in MS Translations service
We had a bug in MSTranslations service. If the translations were made where in the source language was explicitly specified, we would get and exception. It was just a just matter of no null check for the the detected language. We have removed the line for getting detected language, as it was just used for logging on console.